fix: Fix --sign-promo warning

+++ src/TcpConnectionTransport.cpp
@@ -28,7 +28,7 @@ void TcpConnectionTransport::initializeIdentifier() {
     if (getpeername(m_socket, reinterpret_cast<struct sockaddr *>(&addr), &addrLen) == 0) {
         char ipStr[INET_ADDRSTRLEN];
         inet_ntop(AF_INET, &addr.sin_addr, ipStr, sizeof(ipStr));
-        m_identifier = std::string(ipStr) + ":" + std::to_string(ntohs(addr.sin_port));
+        m_identifier = std::string(ipStr) + ":" + std::to_string(static_cast<unsigned int>(ntohs(addr.sin_port)));
     } else {
         m_identifier = "socket_" + std::to_string(m_socket);
     }
